requirements.txt
gradio>=4.26.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
rembg>=2.0.50
trimesh>=4.0.0
einops>=0.7.0
omegaconf>=2.3.0
transformers>=4.38.0
huggingface_hub>=0.21.0
# TripoSR — install from source:
# pip install git+https://github.com/VAST-AI-Research/TripoSR.git
